+++ lib/constants.ts
@@ -3,3 +3,5 @@ export const BLOG_NAME = "piju kafe";
 export const BLOG_PUNCHLINE = "bez mlíka, bez cukru";
 export const KNOW_MORE = "Chcete vědět víc?";
 export const I_WANNA_KNOW_MORE = "Chci vědět víc";
+export const ALL_ARTICLES = "Všechny články";
+export const OLDER_ARTICLES = "Starší články";